+++ src/refactoring/components/cart/CartPage.tsx
@@ -1,4 +1,4 @@
-import { CartItem as CartItemType, Coupon, Product } from "../../../types.ts";
+import { Coupon, Product } from "../../../types.ts";
 import { useCart } from "../../hooks/index.ts";
 import { OrderSummary } from "../order/OrderSummary.tsx";
 import { CouponSelector } from "../coupon/CouponSelector.tsx";
@@ -20,31 +20,13 @@ export const CartPage = ({ products, coupons }: Props) => {
     applyCoupon,
     calculateTotal,
     selectedCoupon,
+    getMaxDiscount,
+    getRemainingStock,
+    getAppliedDiscount,
   } = useCart();
-
-  const getMaxDiscount = (discounts: { quantity: number; rate: number }[]) => {
-    return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
-  };
-
-  const getRemainingStock = (product: Product) => {
-    const cartItem = cart.find((item) => item.product.id === product.id);
-    return product.stock - (cartItem?.quantity || 0);
-  };
 
   const { totalBeforeDiscount, totalAfterDiscount, totalDiscount } =
     calculateTotal();
-
-  const getAppliedDiscount = (item: CartItemType) => {
-    const { discounts } = item.product;
-    const { quantity } = item;
-    let appliedDiscount = 0;
-    for (const discount of discounts) {
-      if (quantity >= discount.quantity) {
-        appliedDiscount = Math.max(appliedDiscount, discount.rate);
-      }
-    }
-    return appliedDiscount;
-  };
 
   return (
     <div className="container mx-auto p-4">
