+++ src/refactoring/hooks/useCart.ts
@@ -40,6 +40,27 @@ export const useCart = () => {
 
   const calculateTotal = () => calculateCartTotal(cart, selectedCoupon);
 
+  const getMaxDiscount = (discounts: { quantity: number; rate: number }[]) => {
+    return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
+  };
+
+  const getRemainingStock = (product: Product) => {
+    const cartItem = cart.find((item) => item.product.id === product.id);
+    return product.stock - (cartItem?.quantity || 0);
+  };
+
+  const getAppliedDiscount = (item: CartItem) => {
+    const { discounts } = item.product;
+    const { quantity } = item;
+    let appliedDiscount = 0;
+    for (const discount of discounts) {
+      if (quantity >= discount.quantity) {
+        appliedDiscount = Math.max(appliedDiscount, discount.rate);
+      }
+    }
+    return appliedDiscount;
+  };
+
   return {
     cart,
     addToCart,
@@ -48,5 +69,8 @@ export const useCart = () => {
     applyCoupon,
     calculateTotal,
     selectedCoupon,
+    getMaxDiscount,
+    getRemainingStock,
+    getAppliedDiscount,
   };
 };
